# openapi.yaml
openapi: 3.0.3
info:
  title: Coachwise API
  version: "0.1.0"
  description: >
    Early-stage REST API for Coachwise. All non-auth routes require
    `Authorization: Bearer <access_token>`.
servers:
  - url: http://localhost:3000
paths:
...
  /workouts/sessions:
    get:
      summary: List user sessions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'
    post:
      summary: Create workout session
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionForm'
      responses:
        "200":
          description: Created session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
  /workouts/sessions/active:
    get:
      summary: List active sessions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Active sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'
  /workouts/sessions/{id}:
    get:
      summary: Get session
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
    put:
      summary: Update session
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSessionForm'
      responses:
        "200":
          description: Updated session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
  /workouts/sessions/{id}/logs:
    get:
      summary: Get session logs
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkoutLog'
  /workouts/logs:
    post:
      summary: Create workout log
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkoutLogForm'
      responses:
        "200":
          description: Created log
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutLog'
  /workouts/logs/{id}:
    put:
      summary: Update workout log
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkoutLogForm'
      responses:
        "200":
          description: Updated log
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutLog'
    delete:
      summary: Delete workout log (not implemented; placeholder response)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Placeholder delete response
  /workouts/logs/{id}/tags:
    post:
      summary: Add tag to workout log
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tag_id:
                  type: string
                  format: uuid
      responses:
        "200":
          description: Tag added
  /tags/search:
    get:
      summary: Search tags
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: query
          schema:
            type: string
        - in: query
          name: sport_type
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: offset
          schema:
            type: integer
      responses:
        "200":
          description: Tags with count
          content:
            application/json:
              schema:
                type: object
                properties:
                  tags:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tag'
                  total_count:
                    type: integer
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Session:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        session_type:
          type: string
          enum: [PLAN, FREESTYLE, STRENGTH, CLIMBING]
        plan_id:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateSessionForm:
      type: object
      required: [session_type]
      properties:
        session_type:
          type: string
          enum: [PLAN, FREESTYLE, STRENGTH, CLIMBING]
        plan_id:
          type: string
          format: uuid
        notes:
          type: string
    UpdateSessionForm:
      type: object
      properties:
        status:
          type: string
          enum: [ACTIVE, COMPLETED, ABANDONED]
        notes:
          type: string
    WorkoutLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        session_id:
          type: string
          format: uuid
        exercise_id:
          type: string
          format: uuid
          nullable: true
        exercise_name:
          type: string
          nullable: true
        set_number:
          type: integer
        reps:
          type: integer
          nullable: true
        weight:
          type: number
          format: float
          nullable: true
        rpe:
          type: number
          format: float
          nullable: true
        duration_seconds:
          type: integer
          nullable: true
        grade:
          type: string
          nullable: true
        completed:
          type: boolean
        attempts:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
    CreateWorkoutLogForm:
      type: object
      required: [session_id, set_number]
      properties:
        session_id:
          type: string
          format: uuid
        exercise_id:
          type: string
          format: uuid
        exercise_name:
          type: string
        set_number:
          type: integer
        reps:
          type: integer
        weight:
          type: number
        rpe:
          type: number
        duration_seconds:
          type: integer
        grade:
          type: string
        completed:
          type: boolean
        attempts:
          type: integer
        notes:
          type: string
        tags:
          type: array
          items:
            type: string
    UpdateWorkoutLogForm:
      type: object
      properties:
        reps:
          type: integer
        weight:
          type: number
        rpe:
          type: number
        duration_seconds:
          type: integer
        grade:
          type: string
        completed:
          type: boolean
        attempts:
          type: integer
        notes:
          type: string
    Tag:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        normalized_name:
          type: string
        is_system:
          type: boolean
        created_at:
          type: string
          format: date-time
...